Updating webserver file structure, using imagemagick to get last display. Adding some admin files used on autotesting.livecd.org - ready for local builds

+++ run-batch-autotesting.sh
@@ -36,23 +36,24 @@
 
 VIDEO_QEMU_BOOTING="/home/autotesting/debian-live/autotesting/video-qemu-booting-iso.sh"
 
-if [ -z "$1" -a -z "$2" -a -z "$3" ]
+if [ -z "$1" -a -z "$2" -a -z "$3" -a -z "$4" ]
 then
     echo "usage: $0 directory/ "
     echo
-    echo " e.g: run-batch-autotesting.sh dailylist.txt iso/daily/ video/daily/"
+    echo " e.g: run-batch-autotesting.sh dailylist.txt iso/daily/ video/daily/ debian-live/i386/ MD5SUM-FILE"
     exit
 fi
 
 URLLIST=$1
 DIRECTORY=$2
 VIDEO_DIRECTORY=$3
+WEB_ROOT=$4
 
-if [ -z "$4" ]
+if [ -z "$5" ]
 then
     MD5SUM_FILE="MD5SUMS"
 else
-    MD5SUM_FILE="$4"
+    MD5SUM_FILE="$5"
 fi
 
 if [ -f /tmp/run-batch-autotesting.lock ]
@@ -75,13 +76,21 @@ do
         echo "$BASE_NAME already downloaded today "
         echo " - skipping download and testing. To retest please remove file or set ctime>1 day"
     else
-        echo "Downloading $URL"
-        rm $DIRECTORY/$BASE_NAME 2>/dev/null
-        rm $DIRECTORY/$MD5SUM_FILE 2>/dev/null
-        wget --no-verbose --tries=3 --timeout=60 --directory-prefix=$DIRECTORY $URL
-        wget --no-verbose --tries=3 --timeout=60 --directory-prefix=$DIRECTORY $MD5SUMS
-        MD5SUM_LOCAL=$(md5sum $DIRECTORY/$BASE_NAME | cut --fields=1 --delimiter=\  )
-        MD5SUM_REMOTE=$(cat $DIRECTORY/$MD5SUM_FILE | grep "$BASE_NAME" | head -n 1 | cut --fields=1 --delimiter=\  )
+        HTTP=${URL:0:4}
+        if [ "$HTTP" = "http" ]
+        then
+            echo "Downloading $URL"
+            rm $DIRECTORY/$BASE_NAME 2>/dev/null
+            rm $DIRECTORY/$MD5SUM_FILE 2>/dev/null
+            wget --no-verbose --tries=3 --timeout=60 --directory-prefix=$DIRECTORY $URL
+            wget --no-verbose --tries=3 --timeout=60 --directory-prefix=$DIRECTORY $MD5SUMS
+            MD5SUM_LOCAL=$(md5sum $DIRECTORY/$BASE_NAME | cut --fields=1 --delimiter=\  )
+            MD5SUM_REMOTE=$(cat $DIRECTORY/$MD5SUM_FILE | grep "$BASE_NAME" | head -n 1 | cut --fields=1 --delimiter=\  )
+        else
+            echo "Local file - assuming already in place"
+            MD5SUM_LOCAL="ok"
+            MD5SUM_REMOTE="$MD5SUM_LOCAL"
+        fi
         if [[ $MD5SUM_LOCAL != $MD5SUM_REMOTE ]]
         then
             echo "$BASE_NAME - md5sums different, remote: $MD5SUM_REMOTE, local: $MD5SUM_LOCAL. "
@@ -91,11 +100,33 @@ do
             DATE_DOWNLOADED=$(ls -lc --time-style long-iso $DIRECTORY/$BASE_NAME |tr -s " "|cut -d" " -f6)
             mkdir "$VIDEO_DIRECTORY/${BASE_NAME}" 
             VIDEO_NAME="$VIDEO_DIRECTORY/${BASE_NAME}/Built_${DATE_IMAGE_BUILT}_Tested_${DATE_DOWNLOADED}_.ogg"
-            LOG_FILE="$VIDEO_DIRECTORY/${BASE_NAME}/Built_${DATE_IMAGE_BUILT}_Tested_${DATE_DOWNLOADED}_.log"
+            LOG_FILE="$VIDEO_DIRECTORY/${BASE_NAME}/Built_${DATE_IMAGE_BUILT}_Tested_${DATE_DOWNLOADED}_log.txt"
             echo "AutoTesting $BASE_NAME $VIDEO_NAME"
-            $VIDEO_QEMU_BOOTING -g 1024x768  -t 300 -v 5  $DIRECTORY/$BASE_NAME $VIDEO_NAME >$LOG_FILE 2>&1
+            $VIDEO_QEMU_BOOTING -g 1024x768  -t 420 -v 5  $DIRECTORY/$BASE_NAME $VIDEO_NAME >$LOG_FILE 2>&1
             echo "Finished Autotesting $BASE_NAME"
             echo
+            echo "----"
+            echo
+            echo "* Creating symlinks to webserver"
+            LN_S_ROOT="$WEB_ROOT/$BASE_NAME/built_${DATE_IMAGE_BUILT}/Tested_${DATE_DOWNLOADED}"
+            LN_S_CURRENT="$WEB_ROOT/$BASE_NAME/current"
+            # making directories if required.
+            mkdir -p "$LN_S_ROOT"
+            mkdir -p "$LN_S_CURRENT"
+            # get rid of the old current
+            rm "$LN_S_CURRENT/*"
+            ln -s $VIDEO_NAME             "$LN_S_ROOT/video-booting.theora.ogg"
+            ln -s $VIDEO_NAME             "$LN_S_CURRENT/video-booting.theora.ogg"
+            ln -s $VIDEO_NAME.montage.jpg "$LN_S_ROOT/video-booting-montage.jpg"
+            ln -s $VIDEO_NAME.montage.jpg "$LN_S_CURRENT/video-booting-montage.jpg"
+            ln -s $VIDEO_NAME.end.jpg     "$LN_S_ROOT/final-screenshot.jpg"
+            ln -s $VIDEO_NAME.end.jpg     "$LN_S_CURRENT/final-screenshot.jpg"
+            SWF_NAME=$(echo $VIDEO_NAME.swf |  rev | cut -d"/" -f 1 | rev)
+            ln -s $VIDEO_NAME.swf         "$LN_S_ROOT/$SWF_NAME"
+            ln -s $VIDEO_NAME.swf         "$LN_S_CURRENT/$SWF_NAME"
+            HTML_PAGE="$VIDEO_DIRECTORY/${BASE_NAME}/Built_${DATE_IMAGE_BUILT}_Tested_${DATE_DOWNLOADED}__View_swf_Video_.html"
+            ln -s $HTML_PAGE              "$LN_S_ROOT/play-video-booting-swf.html"
+            ln -s $HTML_PAGE              "$LN_S_CURRENT/play-video-booting-swf.html"
             echo "----"
             echo
         fi
